Implementation of box size change at constant dx upon input.
New functions & subroutines in Transform.f90, and clear distinction
between interpolation (change of dx that may or may not be accompanied
by a change in nx/ny/nz) and box size change (change of one or several
out of nx, ny, nz at constant dx).

The number of points in each direction can be changed simultaneously,
and they can be increased or decreased independently from each other,
meaning that nx/ny/nz = 20/20/20 -> nx/ny/nz = 18/20/22 is a possible
change of box size that is handled without complaints. The code puts
wave functions and densities into the new box (the latter was necessary
for continuity of constraints, it seems). Additional points are set to
zero, whereas omitted points simply disappear (changing the norm of the
spatial functions as used in the initial step, of course).

The code stops when attempting to change symmetries and change the box
size other than doubling the dimension in the direction of symmetry
breaking.

The box sizes will only be correctly changed when AllowTransform= .true.
When it is not the code will behave erraticially. Unfortunately I did not
manage (yet) to make the code stop when the data indicate box change at
AllowTransform= .false. (the same problem seems to have been present
already for the case of interpolation, as all consistency checks with
respect to that are commented out).

Modified disgnostic printing in InOut.f90 with additional information
about changes of the coordinate space mesh.