Basket of small modifications
1) cutting the (trivial) inverse Jacobian of the no-LN HFB case in
HFBFindFermiEnergyBroyden of HFB.f90, which seems to enormously stabilize
the routine when pairing is breaking down.

2) subroutines to calculate matrix elements of spin in the HF basis in
Spwf.f90 and SpwfStorage.f90. They are printed now for each state. I hope
this provides insight which combinations in multi-qp-states go down in
energy, and which go up.

3) Moments.f90: the beta_{ell m} and beta_ell of protons and neutrons are
now calculated and printed with the correct scaling factor 1/N and 1/Z,
respectively, instead of 1/A.