HFB.f90 ------- new strategy to select blocked quasiparticles, which mainly affects BlockQuasiParticles(). For explanations of the new strategy, see comments in that routine.
New data in Namelist /Blocking/ : "Filled" and "Partners"

New subroutine DecomposeBlockedQuasiParticles() that prints tables providing
an analysis of the blocked qp(s).

New subroutine HFBoverlapT2(overlapT2) calculating the overlap of HF states
with the time-reversed of the other HF states. (HFB.f90 might not be the
best place for this routine as in the end the routine is used differently
than originally planned such that the organization and indexation of arrays
in HFB-style has no added value, quite on the contrary.)

BlockQuasiParticles() is no longer called during the initialization of
HFBFindFermiEnergyBroyden(). In some cases, this utterly confused the
code about the targeted configuration.

Un-commented the commented-out call of QPalignment(). I'm highly suspicious
that the calculation of observables in that routine is either formally wrong
or, if formally correct, what is calculated is useless in practice. The most
interesting information would be how much a given qp contributes to many-body
expectation values of one-body operators, which is entirely determined by the

Various minor bugfixes: use of DIMAG instead of AIMAG, ...

Pairing.f90
-----------
Use the array provided by HFBoverlapT2(overlapT2) to guess partner state
in the HF basis in the no-pairing limit. These are communicated to Spwf
for later printing in the diagnostics of the HF basis and (if the partner
level is not specified on data) used as parter level for blocking.

Spwf.f90
--------
new integer function GetPairPartner(WaveFunction)

Main.f90
--------
change of print format for angular momenta in the iteration analysis

Moments.f90
-----------
Update of print format of magnetic moments required by recent increase
in the default of maximum moment.