New option to additionally precondition the W_munu potential.
modified files:
- Densities.f90
- InOut.f90
- PotentialMixing.f90

There is a new parameter "PreConW" of the namelist /densit/. The default
value is 0.0_dp. If it is not read from data, the code does not precondition
W (as before). If it is read from data and set to a value different from
0.0_dp, the Wmunu potential will be preconditioned. The option seems to have
no visible effect whatsoever beyond changing numerical noise for converged
states.