Further update of HFB pairing, compiles under ifort and gfortran.
Pairing.f90: minor bugfixes concerning compilation with gfortran,
print formats and existence or not of arrays when calling routines.

HFB.f90:

completely rewritten bisection routine HFBFindFermiEnergyBisection()
for HFB using a more efficient initial bracketing of the first guess
and then Brent's method in a new separate subroutine FermiBrent()
to home in on the Fermi energy eps_F where <N(eps_F)>-N_0 changes
sign. At least as efficient as the Broyden solver and absolutely
stable. printing of various warnings if there are anomalies of the
function <N(eps)>, such as
  1) <N(eps)> is decreasing with increasing eps
  2) <N(eps_F)>-N_0 not being anywhere near zero, meaning that
     <N(eps)> has a discontinuity around eps_F.
Both can happen far from convergence, and the new routine handles
these anomalies with a better chance of return to normality in the
following SCF iteration than the Broyden solver.

For details on the algorithms and the adaptation of Brent's method
from the Numerical recipes to some specificities of an HFB
solver see the detailed comments in both routines.

For LN, the old Bisection routine is continued to be used under
the new name of HFBLNFindFermiEnergyBisection() (note additional
"LN" in the name).