New feature: Harmonic Trap
New module defined in

TrapPotential.f90

Numerous small changes (usually a few lines each time) in

Main.f90
MeanFields.f90
InOut.f90
Energy.f90
Force.f90
Densities.f90
ImaginaryTime.f90
HFB.f90
HartreeFock.f90
Moments.f90

in order to handle zero particle number for one nucleon species, for
adding the trap potential to the mean fields and the energy, diagnostic
printing of trap-related observables, and reading of the trap parameters.

See MV's pdf file with notes for instructions how to use the trap.

Note that the trap potential offers unprecedented possibilities to
benchmark new functionals as one can actually converge systems with
1 or 2 particles and test self-interaction and the "magic formula"
simultaneously for cases where time-even and time-odd terms are
of same size.

Hartree-Fock now has an experimental option to use PickHFConfig in case
of broken parity and for time being is just a quick-and-dirty
reinterpretation of the data read for ev8/cr8-like calculations: only the
numbers for positive parity are used.

There is a bugfix in the initial bracketing in the Brent Solver for the
Fermi energy (a .le. became a .eq. or similar) that caused problems in
complete breakdown of pairing.

Note that the Beta_{ 2} is now printed with the correct factor to match
the definition of the Beta_{ l m}. I'm suspicious that the value for
higher Q_{ l} does not latch any widely-used definition, such that
correcting the Beta_{ l} for higher l does not lead to values that
match the  Beta_{ l m}. This needs further attention when these
are actually used one day.