Bugfixes in Moments.f90 concerning the calculation and printing of magnetic moments:
1) g_S,t factors of protons and neutrons were exchanged

2) change of convention for g_L. Introduced the more frequent
convention where g_L,p = 1 and the factor 2 is part of the
geometrical 2/(L+1) factor.

3) 2/(L+1) factor moved from the "physical value" to the
"orbital" value. This factor results from the gradient
in this term, and has nothing to do with electromagnetism.
This modification does not change anything for the dipole,
but the higher moments become smaller.

4) removal of the (for dimensional reasons evidently wrong)
e^2 factor from the orbital contribution to the "physical
value".

5) because of the wrong e^2 factor and the exchanged g_S,t
factors, the "physical value" as printed was always wrong. For
the dipole moment, the printed spin and orbital moments were
correct, such that the physical value can be reconstructed
by script from existing calculations.

New feature:

Additional printing of the cartesian dipole moment mu_z
of the "physical value", as this is what is tabulated in the
literature. It differs from the spherical M_{10} moment by

mu_z = sqrt((4.0_dp*pi)/3.0_dp) * M_{10}

Note that this value only has a physical interpretation as
approximation to the experimental value when the angular
momentum points into z direction.

Note that mu is defined as the expectation value

mu_z = < J J | mu_z | J J >

of angular-momentum eigenstates orientated such that M = J.

In a deformed axial calculation with sufficiently large beta
("strong coupling") this is always the case except for K=1/2
states of J > 1/2 states, for which the angular momentum
usually points into a perpendicular direction.

In a spherical nucleus, the M=J state has to be blocked in
order to obtain an interpretable value of the magnetic moment.

There is now a table with the three individual contributions
to the cartesian dipole moment, such that a-posteriori
manipulations such as rescaling of g factors via scripting
is possible. For the (almost spherical) 7/2- ground state
of Sc41 I find with SLy5s1 when fixing M to 3.5 via a
cranking constraint the code prints (calculation is not
perfectly converged)

____________________________________________________________
 mu_z  n                                           -0.000274
 mu_z  p(S)                                         2.794173
 mu_z  p(L)                                         2.998370
 mu_z  p                                            5.792543
 mu_z  t                                            5.792269

which are the neutron contribution (spin only), the proton spin
and orbital contribution, the sum of the two proton contributions
and the sum of all contributions.

All values are in units of nuclear magnetons.

The Schmidt value in this J=L+1/2 case is 5.7930 mu_N, the
experimental value according to http://magneticmoments.info/data/
is +5.431(2) mu_N.

Doing the same kind of calculation for the 7/2- state in Ca45,
one finds (calculation is not perfectly converged)
____________________________________________________________
 mu_z  n                                           -1.605276
 mu_z  p(S)                                        -0.011656
 mu_z  p(L)                                         0.002225
 mu_z  p                                           -0.009430
 mu_z  t                                           -1.614706

Here, the Schmidt value is -1.9130 mu_N, and the experimental
value (again according to http://magneticmoments.info/data/)
is about -1.32 mu_N, indicating significant core polarisation,
some of which is possibly even described by the HFB calculation.