Minor, but crucial modification of HFBFindFermiEnergyBisection in HFB.f90:
oldpf was never initialised when using FermiSolver='Bisection', such that
when combining this option with PfSolver= .true. the pfaffian of the
HFB Hamiltonian could actually never be used to determine the number parity
of the HFB vacuum before blocking, meaning in practice that the code was run
as if PfSolver= .false.. This caused some interesting problems in tilted-axis
cranked broken-signature calculations, as the code often generated additional
broken high-K pairs instead of collective rotation. In many cases these high-K
states still had the correct overall number parity of protons and neutrons,
but a different number parity than the targeted one in the individual
isospin-parity blocks.

It seems that after the bugfix it is better to set PfSolver= .false. when
putting a signature-conserving state into a signature-breaking box. Needs
further systematic tests.