modified: HFB.f90 , Pairing.f90 , Main.f90
The recent bisection solver for the HFB Fermi energy is minimally updated to
handle also Lipkin-Nogami. Using this update converges some calculations of
cranked blocked transctinides that the default Broyden solver does not converge
(the usual symptom being code crashes with the message "Diagoncr8 failed"),
presumably because it starts to look for the Fermi energy at wrong places.

Various comments added, some actually concerning MB's previous push of MOCCa.

The modifications in detail:

------------------------------------------------------------------------------

HFB.f90:

minimal modification of the new bisection solver to handle also LN.

The routine calculates the lambda2(it) once before iterating the Fermi energy
from the present pairing fields without further updates during the iterations.
With this, the lambda2 will slowly converge with the HF iterations. As the
lambda2 usually will change only by small amounts anyway, this even might
stabilise the iterations.

Using this new option helps to converge some configurations for which the
Broyden solver systematically failed.

The convergence of the bisection solver is somewhat sensitive to the values
of HFBMix. In some difficult cases (for which the Broyden solver also
fails), values of HFBMix = 0.2 are too small and lead to oscillations of
lambda2 that prevent overall convergence. Values of  HFBMix = 0.5 or even
larger usually stabilise the iterations.

------------------------------------------------------------------------------

Pairing.f90:

overload FindFermiEnergy with HFBFindFermiEnergyBisection also in the LN case.

WR's old bisection solver HFBLNFindFermiEnergyBisection is with that probably
obsolete and could be removed in a future cleaning of the code.

In any event, the old bisection solver is unacceptably slow compared to the
new one.

------------------------------------------------------------------------------

Main.f90:

- additional printing of lambda_2 when LN is active. With this, its convergence
  can be directly followed.
- additional printing of ReJxT when time-reversal is broken and signature is
  conserved. This geatly helps to judge convergence (problems) os such
  calculations.
- print format of Fermi energy changed to 5 digits after the decimal point.

------------------------------------------------------------------------------